Erase all references to gwt reflect projects.
Now that gwt reflection support is included in
our fork of Gwt, we can effectively erase all
references to our incubator version of gwt-reflect.

Note that we keep the modules and the tests in our source
tree, as it is easier for us to debug from xapi than gwt,
but we do unhook those modules from their parent pom,
thus making them effectively invisible unless you manually
import the projects and work on them.