Add ThreadLocal scoping to ModelServiceJre
In order for a JRE model service to understand various versions of
de/serialization protocol from clients, it must be able to isolate the version
of the model's being processed from its own runtime classes.

This is done with a ThreadLocal containing a ModelModule that describes
the version of models being used.  This ensures that the server is processing
the version of models the client was compiled with, rather than the server.