Update type signature of X_Collect to be easier (but more dangerou) to use.
Specifically, instead of forcing you to use Class.class.cast(List.class) in
order to get a "Class<List<Foo>>" class instance, the generic signature
is changed to allow any supertype of the expected return type.

This can be dangerous if you try to create an array of the supertype,
as it will then be cast to the subtype and break.  Gwt users may be able
to avoid this by compiling out class cast exceptions altogether (danger!)