Hack in a trick to set style on document fragments
When a potential node wants to insert siblings,
it can opt to use a tagname of "" to create a
document fragment.

The current (bad) api causes this automatic bookkeeping
to lose track of the fact that the author actually
expects a single element, and that it should be
possible to set attributes on that element.

The whole PotentialNode framework should be
viewed as very much a prototype.