Add support for injecting css into shadow dom
Because shadow DOM piercing css selectors are deprecated
and no good, blessed solution to replace it is present
in the shadow DOM spec, we will create our own solution.

Rather than blindly insert all compiled gss into
each web component, it is now possible to specify a
ClientBundle interface, and limit it to include only
select CssResource interface members.

These css resources will be transformed into a style tag,
and that style tag is cloned and inserted into the
shadow DOM.  This (presumably) saves time re-parsing css,
and enables using gss variables/macros instead of raw css.

In the future, as the shadow DOM template system grows,
we will be able to detect which classnames were used,
so we can (possibly) cut the scope of what css is included
to only use those classes that are referenced.