Implement javac-based magic method injection
This rather large commit is the result of a
research project into magic method injection.

We use a java 8 compiler plugin combined with a
fork of the javacc java grammar parser to implement
a templating language in java.  This allows the use
of `string with unescaped newlines
that avoids annoying java "quote escaping" annoyances`,
which are transformed into calls to the magic method
X_Template.parseTemplate().  This method accepts a string,
which it expects to be a compile-time literal,
and returns an object of any type (it is generic).

The magic method compiler plugin finds calls to this
method, checks what type they are being assigned to,
and then chooses the appropriate generator to replace
the given method call with new source code.

There are many egregious hacks that were required
for this prototype; a proper review will be conducted
once the main xapi research project prototype is complete.