Create stub abstractions of a UiElement
Includes a @Ui annotation, a builder system,
and the concept of UiFeature, which is what
we call the "attributes" of "html" elements in XApi.

The long term goal is for developers to be able
to add features across multiple elements by simply
specifying which feature (and feature sets) those
elements support.

Elements with feature support will allow those
features to participate in both compile time and
run time processing of the ui element.

Yes, that last paragraph is extremely generic;
the exact hooks into the build are not yet
well defined, so we will start (in future commits)
to create some default features, like css/style,
id, title/label, src/href, etc. and then use the
lessons learned in supporting those features
to create a more generic, pluggable abstraction layer.