UiElement generic refactor part 2
The base UiElement interface needs to know three types:
The root node type,
The implementors actual element type (extends root type)
And the base UiElement type.

This allows us to create an AbstractUiElement which is
able to implement generic, shared functionality,
then each platform can create a UiElement subclass that
only needs to know the platform-specific element type.

This gives room for abstractions that work across all
platforms, as well as a place for platform-specific
abstractions that can be shared with final implementors.

The generic burden for new element types is a little higher,
but this ensures type safety across all operations (yay!)