Fix TimeService now() and nowPlusOne() methods.
Previous implementation did not consider floating point
data loss when "incrementing" the double timestamp.
In the low precision implementation we had been enforcing
a call to .tick() in order to update the value returned
by .now(), which is confusing to say the least,
and we were adding the much-too-small value of
TimeService.nano_to_second to a millisecond timestamp,
which will just get rounded off (as there are only 0x1000
spaces between floating point timestamps).

Note that there are 41 bits in unix timestamps until 2039.

The high precision implementation was likewise busted, and
both have been changed to use the Math.ulp(timestamp) increment
to ensure that every double returned from .nowPlusOne() is
at least as large as the current timestamp, and always larger
than the previous value returned.

now() has been updated to always return the current system time,
with the lastTick() method added for platforms that wish
to have a stable timestamp (games, for instance, may want to
use tick() per ui frame, then use .lastTick() to ensure stable
for the lifetime of a "single frame" of execution.