Big improvements to ui generator; added <for />
UiTagGenerator is beginning to (has) overflow(n);
in order to break it down into child visitors,
the concept of a visit mode was added to component generator.

This will allow us to create general purpose generators
which will know how to respond, for example,
when they are used in a block of gui definition,
or as part of special tags, like define-tag, if, for, etc.