Add os.detected.classifier as a system-level maven prop
This property can be added to the maven system (and is used by
netty, which we wind up pulling onto classpath when we want
to compile gwt with vertx support).

We add this to our manual hardcoded properties, since it does not
actually appear in parent poms as a property when used.