Add Elemental2 and use it to build virtual shadow DOM
Starts using Elemental2 for better to-the-metal JS support.

We use this to implement virtual shadow DOM, which effectively
injects the shadow DOM directly into the element, then manages
insertions and set-innerHTML correctly.

Insertions cause mutation observers to slot the elements added,
while innerHTML causes the shadow DOM to be temporarily removed,
then the real innerHTML is called, then we put back the shadow DOM
and finally slot all the new elements that were created.

This technique will also enable us to create V0-style content slots
which can use css selectors instead of annoying "set slot= attribute"
from the V1 shadow dom implementation.

Finally, we also add a new web component callback, afterInit,
which is called as soon as possible after the element is contructed,
but after we have escaped the constructor, so that we are allowed
to modify / view children and attributes.