Add ResponseLike, with a default implementation
In an effort to keep "client-side request" separate from
"server-side response", we are adding an ugly extra set
of generics to a RequestScope instance.

If this is too burdensome, we might melt RequestLike and ResponseLike
classes into a single monolithic object (like how vert.x does it),
but, for now, this allows a separation of concerns,
so we can iterate on response building without worrying about
how a given request is received.