Adds CssProxy and enhances component generator
CssProxy is used to read in .cssmap files produced by gwt.
It allows you to create instances of CssResource via reflection Proxy.

This allows server-side rendering to pick out the correct classnames.
Strong-name support via a header is not yet supported;
it will require building a strongname-to-install dir service,
which is not (yet) necessary. For now, each compiled module
will only be able to expose the most recent proxy instance
(in case install reuses output directory, we use a file watcher,
and if file location changes, we have to be notified of the
new install dir (whose contents we then watch as well).

These proxies are relatively expensive, so they should,
ideally, be distributed from a cached resource,
so they can be reused.