Big overhaul of component generator framework
In preparation for demo release, this unfortunately large commit
contains a number of "quick, just make it work" patches, along with
other more well-thought out improvements to component generators.

In particular, this iteration is capable of fully generating
and compiling a set of web components, as well as generating
the xapi server GWTC configuration needed to test said components.

Implementation commit coming soon!