Huge refactor to GwtcService
Detangles the ugly mess that was combining both the generation of
gwt.xml and project files from actually running the compiler.

Adds a new isolated-classloader-safe, inter-process-compatible
monitor class, to allow the calling thread to initiate a Gwt
compilation using a custom classpath in a safe and isolated manner.

This commit includes some half-finished efforts to clean up the
compiler service that will be deleted in the next major commit;
this is "just good enough to pass tests and still work",
the next iteration will be fully polished and ready to show off. :-)