Add MavenLoader for accessing X_Maven without it on the classpath
Ugly as it is, we need to be able to do maven-y things without polluting
the runtime classpath (it has many, many dependencies).  So, when it is
not on the classpath, we will load it, first trying from the local repo,
and downloading from maven central if it is found missing.

We then run it from an isolated classloader, to prevent pollution.