Big upgrade to Gwt Compiler Service
Consolidate a lot of mess, with a clear API for each part of gwtc:
GwtcJobMonitor is responsible for monitoring and communicating with
compilations running in isolated classloaders or new java processes.
GwtcJobManager is responsible for starting jobs and connecting monitors.
GwtcService is now (mostly) stateless, only storing mappings from
module names to job managers (or project generators).
GwtcProjectGenerator is responsible for generating gwt.xml and
java files corresponding to @Gwtc annotations found on a given type.

A bunch of deprecated code is still in existence, and will be deleted
on a future commit.