Make X_Inject handle isolated classloaders
We had previously been using Class.forName, as we had originally
intended for runtime injection to work in gwt clients,
but this has been long since abandoned, so we instead make
the injector prefer thread context classloader.

We also check for parent-child relationships between a
requested class and the context classloader, so we will always
run with the most derived classloader available to us
(otherwise we can miss "patching classloaders" that add
a few urls to another classloader, but we certainly want to scan).