Refactor maven loader + dynamic url handler
This makes apis like gwtc closer to being gwt compatible,
so our clients can treat the server as a simple remote process
and thus use everything a java app can to build a given project.

Also adds "dynamic:$ls" which is a "magic file" that lists all the files
known to the given dynamic url handler.