Gwtc 2.0 - Complete, configurable toolchain for running / serving gwtc
Includes all of the ugly, very-soon-to-be-deleted 1.0 implementations.

Rather than calling different methods to do a given compile
in an ad-hoc manner, we first prepare a complete Gwt project
(gwt.xml, program args, system properties, etc) in the calling thread
then we launch the actual gwt compiler either in an isolated thread,
or in its own process, communicating with a string-based protocol.
(shared classloader is still available, as the default, for cases when
you are supplying classpaths through your own build tool,
but most implementations should prefer a clean room to work in).

Once compilation is complete, the process will stay alive,
listening for tasks (like "die" to shutdown, "recompile", etc).

One bug is that the external processes are actually outliving
the jvm that creates them, so we need to tighten up our cleanup.
A WeakReference<GwtcService> might work, so we can tell when
the client compiler is discarded, and ensure the process goes with it.