Add support for maven resolution of local workspace projects
When adding maven artifacts to a build (like gwtc),
it can be better to point to local source + output,
so you can iterate without rebuilding any jars.

This enhancement will enable this translation,
assuming standard maven source / resource locations are used.

If build-helper-maven-plugin (or other tools) are used,
we may need to consider a way to actually load the project reactor,
instead of merely parsing poms and pulling out values.