Fix changing-order-of-methods causing a bug in jre model service
For JRE implementations, we now read the model fields from bytecode,
where they are (currently) defined in the order they appear in source.

Should this change, we can add a preliminary check for available source
which we can parse, and get 100% reliable method ordering.

This not done now, as type matching is easier in bytecode than source.