Better detection of annotations in xapi-lang
So, looking directly at the underlying char stream is dangerous;
if a previous lookahead already parsed the next token,
you need to check for your expected value there first / as well.

While this commit fixes the bug that was noticed,
it is entirely possible that a similar class of bug exists elsewhere.