Upgrade ui generator based on usage in WTI project
This is the final checkpoint of the hacky prototype for ui generator.
The next phase of development will formally split up initialization
of element structure (creating elements) from rendering elements
(laying out elements).  This will allow us to do a few things:
1) safely defer initialization (provided just-in-time resolution
is added to any exposed means of accessing the component's elements)
2) logically separate "run once" initialization from "run as needed"
layout, which might be triggered multiple times.

We do not want to go the way of react js (throw out all the elements
whenever you detect that you need to change something); rather, we
want to create and update a single set of elements, and just wire
up event handlers / state tracking to know when to re-render some
or all parts of any given component.