Preparing UiTagUiGenerator for massive overhaul.
Replacing giant spaghetti mess of a visitor with a proper object graph;
introduces the AssembledUi object, which will contain per-component
metadata specifically regarding the ui elements (either ui=<dom />,
ui=[<d/>, <o/>, <m/>], or shadow=<dom />).

The notion of shadow dom for non-web components is not fully baked,
so do not use it if you want to support native mobile any time soon.
The plan is (loosely) to simply wire in the notion of slotting,
so we can have "virtual shadow DOM" that is the same in all platforms,
and then get the additional encapsulation on web platforms.

This is a checkpoint commit before deleting current implementation.