Upgrade vert.x xapi server, using vert.x-web Routers
Also upgrades / makes scope api a bit more usable.
This is a checkpoint before a major de-uglification of scope api;

the new API will try to do a bit less, and we might try to extract
the no-vert.x-web version where xapi code handles the complete
abstraction layer, but... for now, we are just going to tightly
couple with what vert.x can get us for free,
and worry about targetting other platforms a) when there is time,
and/or b) when there is a need to make a non-vert.x server.