Huge refactor of ui generator framework
Introduces yet another layer of assembly tools,
in an attempt to further prune past layers of legacy glue.

This time, after experience with the past mess of a visitor,
we are creating abstractions for the players in an assembly
of ui elements, with a UiAssembler responsible for assembling
the whole ui, and TagAssembler, for assembling individual tags.

This produces a much cleaner surface area,
where we can handle things like conditionals and loops
with all assorted complexity living in a single file,
which the assembler calls whenever it needs to create
an appropriate AssembledElement.

In a future commit, we will delete even more of the legacy code,
in particular, ContainerMetadata and ComponentBuffer are both
on the chopping block; ContainerMetadata may get refactored into
"general purpose layer to attach metadata", so we stick with the
current semantics of "first, do a complete visit to check for
metadata about the types of things we have to do, then do another
visit where we actually commit the actions queued up in the first visit.