Huge improvements to web component generator framework
Gets us much closer to where we need to be;
each element in a generated ui is represented by an
AssembledElement, which is stitched together with a TagAssembler,
as orchestrated by a master UiAssembler.

Types can register custom assemblers, which will eventually
be generated (as well as available JustInTime when an element
being included in a ui is currently being generated as well).

Adds native elements, conditionals, ref exposure, builders
that actually do something, and better model binding.