Add support for cross-component builderized-binding, pt 1
Uses ComponentOptions to store a NodeBuilder that can be
used as the element source, so we can create everything
but the element, and have a builder for callers to be
able to plug in our future element.

This commit is a little half baked; the next one will have
a bunch of hackarounds that it took to get to work,
and then a future commit will come back in behind,
and cleanup the spaghetti.