Make generated conditionals insert elements in correct location
Also adds onBeforeCreated() callbacks to ElementBuilder,
for cases like model bindings, that we want to defer,
but which we want to be done before elements are created
(to allow said elements to find and access a primed model).

Also attaches AssembledElement onto the ast node for
said elements (for now it's a single item; might make it a list later).
This is used to do sibling lookups at compile time
(we will traverse the ast and find post-assembled elements to consider).
This is necessary for conditionals to find the correct insertion point.