Add first code generator generator, UiTagAssemblerGenerator!!
Includes various other ui generator improvements as well,
and *finally* generates assembler classes, to automate binding
of components as children of other components.

Note that we'll still need a runtime-only solution,
for cases where generated code is not checked in.
While developing with checked in generated source, however,
simply set `assembler=true` to get an assembler generated,
which you can set as a tag factory to get cross-component binding.

In the future, we'll be nice and scan the directory
of assemblers for you automatically. While this scanning
may make it possible to generate the assemblers up front,
then compile and use them immediately, this may not always
be possible, as it gets into really weird / potentially scary timing.