Refactor core logging, added core injection.
In an attempt to make xapi-fu module self-sufficient,
added thin injection API that searches classpath for
a named file (w/ defaults for singletons|instances).

It does not have the same power and semantics as X_Inject,
but it is thin enough to fit in one file of < 200 lines
(where many lines are documentation, assertions and logging).

This can be used for core types that come before xapi-core-inject
(like our LogInjector).  While the simpler injection semantics of
"first one on classpath wins" is nightmarish with a classpath,
on a modulepath, you have to resolve duplication before runtime
(as you should!), so we'll move in this direction,
and actually use some tools to generate these in transient locations.