Refactor ui generator framework into more general "api generator".
For now, we've moved a bunch of the core basics needed into
the core-lang project, as it was a handy "upstream" to put it in.

These should, however, go somewhere else like xapi-gen;
for now, it is sufficient to break the types out,
to try to escape the plethora of Ui* classes needed for ui gen.

Implements a new tag, <define-api />, for creating models
that are designed to run at the API level (on the wire),
and adding general purpose "applies to whole app",
or "exposes something for all components".

For now, the generated "god factory" is not actually used,
but we can expand upon it as it makes sense to do so.