Absolutely hideous additions to core reflect apis.
Attempted to expand IsClass to sanely handle parameterized types;
wound up trying to build the entire java type system in abstraction,
glued on top of the already-mess we started with.

The result is ...ugly.  Halfway through I gave up and read the whole jls
wept tears of blood, jammed in enough support to make the generator
requiring parameterized types sane enough to use, and swept the rest
"under the rug" (i.e. buried the hacks in this giant commit).

Eventually, when we come back around to it, review the 16 type contexts
and other structures outlined in the jls http://bit.ly/jls_4