Giant ball of code incurred while migrating to gradle
In a perfect world, this would probably be about 12 commits,
or would at least have a commit message describing the many things done.

But, the whole reason it's a ball of code is because we're rushing
to release 0.5.1, and want to get the gradle migration pushed,
so we can begin a whole bunch of move-many-files refactoring commits.

This commit probably does not build in maven anymore;
all of the gradle glue is coming up next.

Not all the gradle tests pass yet,
but the remaining issues are better fixed
by actually building a 1.0-level gradle build tool
(instead of just hack-patching things-that-depend-on-maven).