Migrate to Gradle!
Pt. 1 of the migration: poms still exist, but are basically broken w/out
a primed mavenLocal cache.  Who cares.

auto-migrated gradle version is capable of fully compositing w/ both
collide and wti, and is enough to turn on the prod demo server.

Not all tests pass, but the path to making them pass
is to finally realize the full xapi build chain,
which includes pretty heavy refactors (before and after this commit).