Start of mass-refactor / pom deletion / multi-platform gradle plugin building
For now, we're separating the gradle modules into two layers.
net.wti.gradle.tools contains zero xapi dependencies,
and defines the basic tools that we want to apply to all projects.
net.wti.gradle will likely become net.wti.gradle.plugin;
it depends on xapi-lang, so it can parse supplied manifests.

We also started building the multi-sourceset inject/collect projects,
with just light ripping and tearing to make get gradle wiring in place.

As of this checkpoint, the command line and IDE both build,
but GWT tests and a few others are still failing;
this will only be remedied after we finish the gwt-platform wiring,
so we can delete the hideous hacks we've made due with until now.