Experimental new xapi-schema plugin; likely to deprecate old plugin.
The current incarnation of our xapi plugin exposes the dsl we want.

It is meant to give a concise, powerful, accurate configuration syntax
to describe multiple groups of sourcesets belonging to a single project.

An ArchiveConfig describes a single source set per platform,
and any archives they might require (i.e. main requires api and spi).
A PlatformConfig describes a group of ArchiveConfigs.

There is a "root" main (and test) PlatformConfig, which are
bound to the existing, familiar java plugin sourcesets.
Any additional platforms must replace() either of the roots.

When a platform replaces another, it creates a new layer
of ArchiveConfig which extend the parent platform.  In the future,
a selector function will also be exposed, to allow configuring
when a given sub-platform is chosen.

In a future commit, we will add an `impl` archive type
which will use configuration attributes to setup proper gradle variants.

This will help enforce correct dependency chains, as any impl-platform
(non main/test/root) should not be able to import code from any other
impl-platform (that it doesn't directly replace).

More information is given in the javadocs of the committed code.