Add XapiRequire plugin, and prototype of xapi BuildGraph
This is a checkpoint of a (sort of) working xapi-schema plugin.
The next step will be to make the use of xapiSchema {} very rare,
and instead build up per-project BuildGraph of "actual sources /
configurations / tasks" that we are going to operate on.

The XapiSchemaPlugin is responsible for registering platform/archive
definitions in the BuildGraph, while XapiRequirePlugin is responsible
for realizing some/all of that registered build graph.