Big chunk of work to get inter-project dependencies working
While we can now depend on configurations from external projects,
we've broken depending on configurations from within a given project.

This breakage is due to a need for disambiguation rules
(or, possibly, saner / cleaner Configuration semantics).

We'll get that fixed, and tests all passing on the next commit;
we just want a checkpoint here since the next round of work may
need to be rolled back if it doesn't go so well...