Add XapiJavaPlugin, to replace / avoid stock java plugin
The basic java plugin adds a mess of configurations that we don't want.
While we do want to be able to survive mixing with the java plugin,
we are internally going to avoid it entirely, so we don't have a ton
of unused configurations in play.

We may regret / reverse this decision later, but for now, we will
maintain a separate plugin to use when the user has not either added
java / java-library plugins (or explicitly request that we add them).