Completely remove java plugin from main xapi project
Attempting to disambiguate against changing java plugin
setup is causing very-difficult-to-workaround problems
(java's apiElements variant conflicting with xapi exportCompile).

As a temporary means to be able to at least finish porting
the hideous from-maven module structure, we are simply
removing the use of java plugin until all cleanup
and refactoring is complete.  At that time, we will revisit
supporting full interop w/ the upstream java plugin,
as we can expect the plugin code to have matured a lot by then.

This also moves xapi-core-api into xapi-common:main:main;
in a future commit, we will also start moving more and more
code into consolidated projects, and start deleting
the leftover sprawl-from-maven.

Note that, at this time, xapiRequire desperately needs
a few improvements: The ability to select whether an
external dependency should be treated as "built-by-xapi",
and the ability to configure transitivity promotion
(`implementation-style` "needed for compilation,
but not transitive until runtime).

There's a lot more cleanup to go, but this builds,
and it allows us to finish fixing the hideous project structure
so we can tend to the logic in a cleaner setup.