Re-enable minimal publishing by tracking liveness of dependency graph
At one point, we could only check if a module had sources
to determine whether to realize it or not.

If a module was just pointing to other dependences with sources,
the necessary configurations were not being created in time,
so we just realized every possible platform + module permutation.

By adding actual support for ArchiveRequest (ModuleRequest, soon)
we track incoming and outgoing dependencies:

Consider ModA and ModB.
ModA.dependsOn ModB

ModA.incoming = [ModB]
ModB.outgoing = [ModA]

Now we check: has source? is used? uses something else?

Currently, if any of these are true for a module, it is realized.

This is certainly imperfect,
but it's enough for the xapi build to survive for now
(with 6s rebuilds instead of 26s).