Beginning of idea for final-form of schema.xapi parser.
Attempts at making code asynchronous keeps leading to spaghetti mess.

The latest attempt w/ xapi-loader is still saveable, and this commit
is the design which will save this model.

This time, we have a clear set of timing rules to obey
about what parses what when, and there will only ever be a single
blocker trying load anything important, so we can't ever deadlock
if somebody tries to fully-resolve-the-whole-world at the wrong time.