Fix xapi lang bug around lookahead-crossing-4096-characters.
There may be other places where this occurs...
It is likely that putBackTokens has the same problem.

A note was left about the bug in the place where it will break in the
future, but I don't have time to hunt a bug that's not harming anything
atm.  Will need to trigger this with a setState around UiBody or
CssContainer.

Nothing under active development uses these structures atm,
so this bug lives another day.